requirements.txt
Flask==3.0.3
moviepy==1.0.3
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.0.1
python-dotenv==1.0.1
